use a wrapper.
