+++ python/graphstorm/gconstruct/construct_graph.py
@@ -31,14 +31,14 @@ import dgl
 
 from ..utils import sys_tracker
 from .file_io import parse_node_file_format, parse_edge_file_format
-from .file_io import get_in_files, HDF5Array
+from .file_io import get_in_files
 from .transform import parse_feat_ops, process_features, preprocess_features
 from .transform import parse_label_ops, process_labels
 from .transform import do_multiprocess_transform
 from .id_map import NoopMap, IdMap, map_node_ids
 from .utils import (multiprocessing_data_read,
                     update_two_phase_feat_ops, ExtMemArrayMerger,
-                    partition_graph)
+                    partition_graph, ExtMemArrayWrapper)
 
 def prepare_node_data(in_file, feat_ops, read_file):
     """ Prepare node data information for data transformation.
@@ -301,9 +301,9 @@ def process_node_data(process_confs, arr_merger, remap_id, num_processes=1):
                 if feat_name not in type_node_data:
                     type_node_data[feat_name] = [None] * len(return_dict)
                 type_node_data[feat_name][i] = data[feat_name]
-            # If it's HDF5Array, it's better to convert it into a Numpy array.
+            # If it's ExtMemArray, it's better to convert it into a Numpy array.
             # This will make the next operations on it more efficiently.
-            if isinstance(node_ids, HDF5Array):
+            if isinstance(node_ids, ExtMemArrayWrapper):
                 type_node_id_map[i] = node_ids.to_numpy()
             else:
                 type_node_id_map[i] = node_ids
@@ -659,13 +659,13 @@ def process_graph(args):
     elif args.output_format == "DGL":
         for ntype in node_data:
             for name, ndata in node_data[ntype].items():
-                if isinstance(ndata, HDF5Array):
+                if isinstance(ndata, ExtMemArrayWrapper):
                     g.nodes[ntype].data[name] = ndata.to_tensor()
                 else:
                     g.nodes[ntype].data[name] = th.tensor(ndata)
         for etype in edge_data:
             for name, edata in edge_data[etype].items():
-                if isinstance(edata, HDF5Array):
+                if isinstance(edata, ExtMemArrayWrapper):
                     g.edges[etype].data[name] = edata.to_tensor()
                 else:
                     g.edges[etype].data[name] = th.tensor(edata)
