+++ python/graphstorm/gconstruct/file_io.py
@@ -27,6 +27,8 @@ import numpy as np
 import torch as th
 import h5py
 
+from .utils import HDF5Handle, HDF5Array
+
 def read_index_json(data_file):
     """ Read the index from a JSON file.
 
@@ -176,119 +178,6 @@ def write_data_parquet(data, data_file):
             arr_dict[key] = [arr[i] for i in range(len(arr))]
     table = pa.Table.from_arrays(list(arr_dict.values()), names=list(arr_dict.keys()))
     pq.write_table(table, data_file)
-
-class HDF5Handle:
-    """ HDF5 file handle
-
-    This is to reference the HDF5 handle and close it when no one
-    uses the HDF5 file.
-
-    Parameters
-    ----------
-    f : HDF5 file handle
-        The handle to access the HDF5 file.
-    """
-    def __init__(self, f):
-        self._f = f
-
-    def __del__(self):
-        return self._f.close()
-
-
-class HDF5Array:
-    """ This is an array wrapper class for HDF5 array.
-
-    The main purpose of this class is to make sure that we can close
-    the HDF5 files when the array is destroyed.
-
-    Parameters
-    ----------
-    arr : HDF5 dataset
-        The array-like object for accessing the HDF5 file.
-    handle : HDF5Handle
-        The handle that references to the opened HDF5 file.
-    """
-    def __init__(self, arr, handle):
-        self._arr = arr
-        self._handle = handle
-        self._out_dtype = None # Use the dtype of self._arr
-
-    def __len__(self):
-        return self._arr.shape[0]
-
-    def __getitem__(self, idx):
-        """ Slicing data from the array.
-
-        Parameters
-        ----------
-        idx : Numpy array or Pytorch tensor or slice or int.
-            The index.
-
-        Returns
-        -------
-        Numpy array : the data from the HDF5 array indexed by `idx`.
-        """
-        if isinstance(idx, (slice, int)):
-            return self._arr[idx].astype(self._out_dtype)
-
-        if isinstance(idx, th.Tensor):
-            idx = idx.numpy()
-        # If the idx are sorted.
-        if np.all(idx[1:] - idx[:-1] > 0):
-            arr = self._arr[idx]
-        else:
-            # There are two cases here: 1) there are duplicated IDs,
-            # 2) the IDs are not sorted. Unique can return unique
-            # IDs in the ascending order that meets the requirement of
-            # HDF5 indexing.
-            uniq_ids, reverse_idx = np.unique(idx, return_inverse=True)
-            arr = self._arr[uniq_ids][reverse_idx]
-
-        if self._out_dtype is not None:
-            arr = arr.astype(self._out_dtype)
-        return arr
-
-    def to_tensor(self):
-        """ Return Pytorch tensor.
-        """
-        arr = th.tensor(self._arr)
-        if self._out_dtype is not None:
-            if self._out_dtype is np.float32:
-                arr = arr.to(th.float32)
-            elif self._out_dtype is np.float16:
-                arr = arr.to(th.float16)
-        return arr
-
-    def to_numpy(self):
-        """ Return Numpy array.
-        """
-        res = self._arr[:]
-        if self._out_dtype is not None:
-            res = res.astype(self._out_dtype)
-        return res
-
-    def astype(self, dtype):
-        """ Set the output dtype.
-
-        Parameters
-        ----------
-        dtype: numpy.dtype
-            Output dtype
-        """
-        self._out_dtype = dtype
-        return self
-
-    @property
-    def shape(self):
-        """ The shape of the HDF5 array.
-        """
-        return self._arr.shape
-
-    @property
-    def dtype(self):
-        """ The data type of the HDF5 array.
-        """
-        return self._arr.dtype
 
 def read_data_hdf5(data_file, data_fields=None, in_mem=True):
     """ Read the data from a HDF5 file.
